wren-ai-service/pyproject.toml
[tool.poetry]
name = "wren-ai-service"
version = "0.29.3"
description = ""
authors = ["dev@getwren.ai"]
license = "AGPL-3.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12.*, <3.13"
fastapi = "^0.121.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"} # Litellm requires uvicorn < 0.30.0
python-dotenv = "^1.0.1"
haystack-ai = "==2.26.0"
openai = "^1.40.0"
qdrant-haystack = "^10.2.1"
backoff = "^2.2.1"
tqdm = "^4.66.4"
numpy = "^1.26.4"
sqlparse = "^0.5.0"
orjson = "^3.11.5"
sf-hamilton = {version = "^1.69.0"}
aiohttp = {extras = ["speedups"], version = "^3.13.3"}
ollama-haystack = "^6.2.0"
langfuse = "^4.0.0"
ollama = "^0.6.1"
toml = "^0.10.2"
cachetools = "^5.5.0"
pyyaml = "^6.0.2"
pydantic-settings = "^2.5.2"
google-auth = "^2.35.0"
tiktoken = "^0.8.0"
jsonschema = "^4.23.0"
litellm = "^1.75.2"
boto3 = "^1.34.34" # Litellm requires boto3 = 1.34.34
qdrant-client = "==1.17.1"
filelock = "^3.20.1"
urllib3 = ">=2.6.3"
werkzeug = ">=3.1.5"  # CVE: safe_join Windows device name vulnerability
marshmallow = ">=3.26.2"  # CVE: Schema.load(many=True) DoS

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
streamlit = "^1.37.0"
watchdog = "^4.0.0"
pandas = "^2.2.2"
matplotlib = "^3.9.2"
sseclient-py = "^1.8.0"
dspy-ai = "^2.5.26"
requests = "^2.32.4"
extra-streamlit-components = "^0.1.71"
deepeval = "^3.9.2"
tomlkit = "^0.13.0"
nltk = "^3.9.1"
psycopg2 = "^2.9.10"

[tool.poetry.group.eval.dependencies]
gitpython = "^3.1.43"
plotly = "^5.24.1"
nbformat = "^5.1.3"
ipykernel = "^6.29.5"
itables = "^2.2.1"
gdown = "^5.2.0"
streamlit-tags = "^1.2.8"
docker = "^7.1.0"

[tool.poetry.group.test.dependencies]
locust = "^2.32.0"
pytest = "^8.3.0"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.24.0"
aioresponses = "^0.7.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
